Move class to cpp
