+++ ydb/core/nbs/cloud/blockstore/libs/service/overlapped_requests_guard_wrapper.h
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <ydb/core/nbs/cloud/blockstore/libs/common/block_range_map.h>
-#include <ydb/core/nbs/cloud/blockstore/libs/service/storage.h>
+#include "public.h"
 
 namespace NYdb::NBS::NBlockStore {
 
@@ -17,42 +16,9 @@ namespace NYdb::NBS::NBlockStore {
 // reordered, and we pretend that the #2 request was executed first, and was
 // then completely rewritten by #1.
 // Read requests are not affected in any way.
-class TOverlappedRequestsGuardStorageWrapper final
-    : public IStorage
-    , public std::enable_shared_from_this<
-          TOverlappedRequestsGuardStorageWrapper>
-{
-private:
-    struct TInflight;
+// Thread-safe.
 
-    const IStoragePtr Storage;
-
-    TAdaptiveLock Lock;
-    ui64 RequestIdGenerator = 0;
-    TBlockRangeMap<ui64, std::unique_ptr<TInflight>> InflightRequests;
-
-public:
-    explicit TOverlappedRequestsGuardStorageWrapper(IStoragePtr storage);
-    ~TOverlappedRequestsGuardStorageWrapper() override;
-
-    // implements IStorage
-    NThreading::TFuture<TReadBlocksLocalResponse> ReadBlocksLocal(
-        TCallContextPtr callContext,
-        std::shared_ptr<TReadBlocksLocalRequest> request) override;
-
-    NThreading::TFuture<TWriteBlocksLocalResponse> WriteBlocksLocal(
-        TCallContextPtr callContext,
-        std::shared_ptr<TWriteBlocksLocalRequest> request) override;
-
-    NThreading::TFuture<TZeroBlocksLocalResponse> ZeroBlocksLocal(
-        TCallContextPtr callContext,
-        std::shared_ptr<TZeroBlocksLocalRequest> request) override;
-
-    void ReportIOError() override;
-
-private:
-    void OnRequestFinished(ui64 requestId, const NProto::TError& error);
-};
+IStoragePtr CreateOverlappedRequestsGuardStorageWrapper(IStoragePtr storage);
 
 ////////////////////////////////////////////////////////////////////////////////
