+++ ydb/core/nbs/cloud/blockstore/libs/service/overlapped_requests_guard_wrapper_ut.cpp
@@ -33,8 +33,7 @@ struct TTestEnvironment
     TTestEnvironment()
     {
         Storage = std::make_shared<TTestStorage>();
-        Wrapper =
-            std::make_shared<TOverlappedRequestsGuardStorageWrapper>(Storage);
+        Wrapper = CreateOverlappedRequestsGuardStorageWrapper(Storage);
 
         Storage->ReadBlocksLocalHandler =
             [&](TCallContextPtr callContext,
